third_party/blink/web_tests/virtual/threaded-prefer-compositing/fast/scroll-behavior/smooth-scroll/main-thread-scrolling-reason-correctness-expected.txt

This test verifies that the ScrollAnimator can schedule animations on the compositor when it adds a temporary main thread scrolling reason.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.getScrollAnimationState(document) is RUNNING_ON_COMPOSITOR
PASS internals.mainThreadScrollingReasons(document) is ANIMATING_TEXT
PASS needsUpdateOrRunningOnCompositor(document) is true
PASS internals.mainThreadScrollingReasons(document) is ANIMATING_TEXT
PASS needsUpdateOrRunningOnCompositor(document) is true
PASS successfullyParsed is true

TEST COMPLETE
